backend/requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
google-generativeai==0.8.4
openai
python-dotenv==1.0.1
sqlalchemy==2.0.38
pydantic==2.10.6
beautifulsoup4==4.13.3
httpx==0.28.1
python-multipart==0.0.20
jinja2==3.1.5
openai
